+++ Source/OCMockito/Core/OCMockito.m
@@ -89,3 +89,10 @@ id MKTAtMost(NSUInteger maxNumberOfInvocations)
 {
     return [[MKTAtMostTimes alloc] initWithMaximumCount:maxNumberOfInvocations];
 }
+
+void MKTStopMockingWithLocation(id mock, id testCase, const char *fileName, int lineNumber)
+{
+    if (reportedInvalidMock(mock, testCase, fileName, lineNumber, @"stopMocking()"))
+        return;
+    [mock mkt_stopMocking];
+}